requirements.txt
streamlit==1.35.0
groq==0.9.0
PyMuPDF==1.24.5
youtube-transcript-api==0.6.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
